save2

+++ src/main/java/com/project/travelAgency/service/TourService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface TourService {
 
     List<TourDto> getAll();
-    void addToUserCart(Integer tourId,String username);// добавляем тур по id к определенному user
+    void addToUserCart(Long tourId,String username);// добавляем тур по id к определенному user
 }
